+++ ui/src/auth/index.js
@@ -1,2 +1,3 @@
 import Login from './Login'
-export {Login}
+import {UserIsAuthenticated, Authenticated, UserIsNotAuthenticated} from './Authenticated'
+export {Login, UserIsAuthenticated, Authenticated, UserIsNotAuthenticated}