+++ ui/src/shared/components/AutoRefresh.js
@@ -69,6 +69,8 @@ const AutoRefresh = (ComposedComponent) => {
       const newSeries = []
       for (const query of queries) {
         const {host, database, rp} = query
+        // TODO: enact this via an action creator so redux will know about it; currently errors are used as responses here
+        // TODO: may need to make this a try/catch
         const response = await fetchTimeSeriesAsync({source: host, db: database, rp, query}, this.props.editQueryStatus)
         newSeries.push({response})
         count += 1