+++ ui/src/side_nav/index.js
@@ -1,2 +1,2 @@
-import SideNavApp from './containers/SideNavApp'
-export default SideNavApp
+import SideNav from 'src/side_nav/containers/SideNav'
+export default SideNav