+++ ui/src/store/configureStore.js
@@ -1,6 +1,9 @@
 import {createStore, applyMiddleware, compose} from 'redux'
 import {combineReducers} from 'redux'
+import {routerReducer, routerMiddleware} from 'react-router-redux'
 import thunkMiddleware from 'redux-thunk'
+
+import errorsMiddleware from 'shared/middleware/errors'
 import resizeLayout from 'src/shared/middleware/resizeLayout'
 import adminReducer from 'src/admin/reducers/admin'
 import sharedReducers from 'src/shared/reducers'
@@ -15,14 +18,16 @@ const rootReducer = combineReducers({
   admin: adminReducer,
   rules: rulesReducer,
   dashboardUI,
+  routing: routerReducer,
 })
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
 
-export default function configureStore(initialState) {
+export default function configureStore(initialState, browserHistory) {
+  const routingMiddleware = routerMiddleware(browserHistory)
   const createPersistentStore = composeEnhancers(
     persistStateEnhancer(),
-    applyMiddleware(thunkMiddleware, resizeLayout),
+    applyMiddleware(thunkMiddleware, routingMiddleware, errorsMiddleware, resizeLayout),
   )(createStore)